+++ server/vroom/src/main/java/org/example/vroom/exceptions/auth/InvalidLoginException.java
@@ -1,4 +1,4 @@
-package org.example.vroom.exceptions;
+package org.example.vroom.exceptions.auth;
 
 public class InvalidLoginException extends RuntimeException{
     public InvalidLoginException(String message){
