+++ server/vroom/src/main/java/org/example/vroom/exceptions/auth/InvalidTokenException.java
@@ -1,4 +1,4 @@
-package org.example.vroom.exceptions;
+package org.example.vroom.exceptions.auth;
 
 public class InvalidTokenException extends RuntimeException{
     public InvalidTokenException(String message){
