+++ server/vroom/src/main/java/org/example/vroom/exceptions/auth/TokenPresentException.java
@@ -1,4 +1,4 @@
-package org.example.vroom.exceptions;
+package org.example.vroom.exceptions.auth;
 
 public class TokenPresentException extends RuntimeException{
     public TokenPresentException(String message){
