+++ server/vroom/src/main/java/org/example/vroom/exceptions/user/AccountStatusException.java
@@ -1,4 +1,4 @@
-package org.example.vroom.exceptions;
+package org.example.vroom.exceptions.user;
 
 public class AccountStatusException extends RuntimeException{
     public AccountStatusException(String message){
