+++ server/vroom/src/main/java/org/example/vroom/exceptions/user/UserAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package org.example.vroom.exceptions;
+package org.example.vroom.exceptions.user;
 
 public class UserAlreadyExistsException extends RuntimeException{
     public UserAlreadyExistsException(String message) {
