+++ server/vroom/src/main/java/org/example/vroom/exceptions/user/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package org.example.vroom.exceptions;
+package org.example.vroom.exceptions.user;
 
 public class UserNotFoundException extends RuntimeException{
     public UserNotFoundException(String message){
